Alekhya | MOBN-984 | Move FocusLock to utils/hooks
